Use temporary files for shards and encrypted file
Use SpooledTemporaryFiles instead of creating real files. This way the
file is spooled in memory up to 10MiB, and is backed on hard disk only
when in pass the 10MiB threshold.

The additional benefit is that we do not need to discover the temp
directory and we do not need to ensure deletion of the file. Everything
is managed by python itself.